+++ io_mesh_roomle/optimize_operator.py
@@ -27,8 +27,8 @@ class OptimizeSceneOperator(bpy.types.Operator):
     bl_label = "Optimize Roomle static"
     bl_options = {'REGISTER', 'UNDO'} # 'PRESET'
 
-    center_scene = BoolProperty(name="Center Scene", description="center scene horicontally and place it on x-y plane vertically.", default=True)
-    reset_transforms = BoolProperty(name="Reset Transforms", description="Remove hierarchy, apply scale/rotation/translation", default=True)
+    center_scene: BoolProperty(name="Center Scene", description="center scene horicontally and place it on x-y plane vertically.", default=True)
+    reset_transforms: BoolProperty(name="Reset Transforms", description="Remove hierarchy, apply scale/rotation/translation", default=True)
 
     @classmethod
     def poll(cls, context):